refactor(api): move findCertificationCenterMembership controller method
to src/team/certification-center-membership
