+++ api/tests/unit/application/certification-centers/certification-center-controller_test.js
@@ -150,48 +150,6 @@ describe('Unit | Controller | certifications-center-controller', function () {
     });
   });
 
-  describe('#findCertificationCenterMemberships', function () {
-    it('should return the serialized membership', async function () {
-      // given
-      const user = domainBuilder.buildUser();
-      const certificationCenter = domainBuilder.buildCertificationCenter();
-      const certificationCenterMembership = domainBuilder.buildCertificationCenterMembership({
-        certificationCenter,
-        user,
-      });
-      const serializedCertificationCenterMembership = Symbol('certification center membership serialized');
-
-      const request = {
-        params: {
-          certificationCenterId: certificationCenter.id,
-        },
-      };
-
-      sinon
-        .stub(teamUsecases, 'findCertificationCenterMembershipsByCertificationCenter')
-        .withArgs({
-          certificationCenterId: certificationCenter.id,
-        })
-        .resolves(certificationCenterMembership);
-
-      const certificationCenterMembershipSerializerStub = {
-        serializeMembers: sinon.stub(),
-      };
-      certificationCenterMembershipSerializerStub.serializeMembers
-        .withArgs(certificationCenterMembership)
-        .returns(serializedCertificationCenterMembership);
-
-      // when
-      const response = await certificationCenterController.findCertificationCenterMemberships(request, hFake, {
-        certificationCenterMembershipSerializer: certificationCenterMembershipSerializerStub,
-      });
-
-      // then
-      expect(teamUsecases.findCertificationCenterMembershipsByCertificationCenter).to.have.been.calledOnce;
-      expect(response).equal(serializedCertificationCenterMembership);
-    });
-  });
-
   describe('#createCertificationCenterMembershipByEmail', function () {
     const certificationCenterId = 1;
     const email = 'user@example.net';
